// src/app/shared/model/strenna.model.ts
export interface Strenna {
    id_strenna?: string;
    period: string;
    sender: string;
    receiver: string;
    sender_city?: string;
    method: string;
    courier?: string;
    lettura_vettura?: string;
    type: string;
    date: string;
    stock_position?: string;
    status: string;
    response: string;
    note: string;
    bulky: string;
    perisheable: string;
}

export interface StrennaDTO{
    id_strenna?: string;
    period: string;
    sender: string;
    receiver: string;
    sender_city?: string;
    method: string;
    courier?: string;
    lettura_vettura?: string;
    type: string;
    date: string;
    stock_position?: string;
    status: string;
    response: string;
    note: string;
    bulky: string;
    perisheable: string;
 } 